+++ CHANGES.md
@@ -2,5 +2,13 @@
 
 Any modifications to the OpenAmiga500FastRamExpansion design and documentation made by any Licensee MUST be noted in this file, as per *section 3.4.b* of the [license](LICENSE.pdf).
 
-### (Date) - (Author)
-(Description)
+### 04/02/2021 - Matt Harlum (LIV2)
+* Moved MA0/1 to another 74157
+* Added _cfgin pin
+* Added autoconfig hiding circuit
+
+### 13/02/2021 - Matt Harlum (LIV2)
+* Added /AS to the autoconfig hiding circuit
+* Add 4MB jumper
+* Add Firmware source and binaries
+* Update documentation for board V2